backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
bleach==6.1.0
requests==2.31.0
psycopg2-binary==2.9.9
stripe==7.8.0
